fix(recover): centralize models + cap limit + docs (#834 Phase 1c)
Addresses PR #840 review (vybe, CHANGES_REQUESTED); rebased onto the
updated #839/#838.

- Move SoftDeletedAgent / SoftDeletedSchedule from routers/admin_recovery.py
  to models.py (Architectural Invariant #14 — response models are
  centralized, not declared in router files).
- Cap limit on both list endpoints: limit: int = Query(200, le=500) —
  prevents unbounded DB reads (422 on >500).
- requirements.md: §32.3 fleshed out (endpoints, recovery semantics,
  metadata-only/container-recreate-is-Phase-2, audit events, model
  location).
- architecture.md: new "Soft-Delete Admin Recovery (#834 Phase 1c)"
  endpoint table with all 4 endpoints, auth, and behavior notes.

Related to #834
